% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cck94v2-package.R
\docType{package}
\name{bw05-package}
\alias{bw05}
\alias{bw05-package}
\title{bw05: Ramsey Optimal Taxation Using a Monte Carlo Linear-Quadratic Approach (Benigno and Woodford, 2005)}
\description{
Replication of Chari, Christiano and Kehoe (1994) Table 2 using the approach from Benigno and Woodford (2005) to solve the Ramsey optimal taxation problem. The code is structured in a way that allows for full replication of Table 5 in BW05, which matches CCK94 results.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
