% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bw05_params}
\alias{bw05_params}
\title{NSSS and LQ diagnostics}
\usage{
bw05_params(args)
}
\description{
Returns steady-state shares, BW utility coefficients, and
  LQ loss function coefficients for diagnostic purposes.
}
