% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{bw05_stats}
\alias{bw05_stats}
\title{BW Table 5: 1st-order Monte Carlo statistics}
\usage{
bw05_stats(args, n_k, T_total, T_burn, seed)
}
\description{
Simulates log-linearized Ramsey policy rules using BW's
  LQ approach and returns the same statistics as Table 5 of Benigno &
  Woodford (2005).
}
